Update edge.h
Changed weight from int to double

+++ src/edge.h
@@ -87,7 +87,7 @@ class Edge
     /**
      * Gets edge weight.
      */
-    int getWeight() const
+    double getWeight() const
     {
         return this->weight;
     }
@@ -106,6 +106,6 @@ class Edge
     }
 private:
     string label; /**< The edge label **/
-    int weight; /**< The edge weight (if in a weighed graph) **/
+    double weight; /**< The edge weight (if in a weighed graph) **/
 
 };
